Refactor HomeScreen component
